feature: changed service to snakecase
