+++ mp-extended/src/main/java/com/mercadopago/sdk/android/coremethods/builder/data/remote/request/MPDeviceSessionIdRequest.kt
@@ -1,8 +1,10 @@
 package com.mercadopago.sdk.android.coremethods.builder.data.remote.request
 
+import com.google.gson.annotations.SerializedName
 import com.mercadolibre.android.device.sdk.domain.Fingerprint
 
 internal data class MPDeviceSessionIdRequest(
-    val device: Fingerprint?,
+    @SerializedName("finger_print")
+    val fingerprint: Fingerprint?,
     val siteId: String,
 )
